+++ modbus/modbus_client.h
@@ -25,7 +25,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
  * @author Oryx Embedded SARL (www.oryx-embedded.com)
- * @version 2.1.6
+ * @version 2.1.8
  **/
 
 #ifndef _MODBUS_CLIENT_H
@@ -166,7 +166,7 @@ error_t modbusClientReadCoils(ModbusClientContext *context,
    uint16_t address, uint_t quantity, uint8_t *value);
 
 error_t modbusClientReadDiscreteInputs(ModbusClientContext *context,
-   uint8_t address, uint_t quantity, uint8_t *value);
+   uint16_t address, uint_t quantity, uint8_t *value);
 
 error_t modbusClientReadHoldingRegs(ModbusClientContext *context,
    uint16_t address, uint_t quantity, uint16_t *value);
